2009-04-22  Manuel Arostegui <marostegui@artica.es>
	* install.php: Fixed bug #2777317

	* godmode/agentes/planned_downtime.php: Fixed bug #2777475

	* godmode/reporting/map_builder.php: Fixed bug #2777347



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@1642 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/install.php
@@ -452,8 +452,8 @@ $config["homeurl"]="'.$url.'";			// Base URL
 
 				if (mysql_error() != "")
 					echo "<div class='warn'> <b>ERROR:</b> ". mysql_error().".</div>";
-
-				mysql_query ("DROP DATABASE $dbname");
+				if ($step1 == 1)
+					mysql_query ("DROP DATABASE $dbname");
 			}		
 		echo "
 		</div>
